Added Simple Material-Based Wind Effects, Gave Common Materials Names, and Improved Title Intro Sequence
Foliage now subtle sways in the wind. This is based in their material and will not adjust to wind in the level. It is a placeholder for a better system down the line.

The old foliage static meshes used materials with generic names. Giving each one a unique name helped simplify the materials list and allows for custom materials per static mesh.

The Intro Sequence is now timed better and slightly longer.