Add AddHealth Function Override in JEnemy_BP, Fix Life Alert Looking for None, Zorrander Smash Now Cancels After A Certain Ammount Of Cumulative Damage.
The TheoChanges version is for Patrick to bring over my new stuff to add to his seperation and unification logic.

INFO:
JEnemy.cpp has a AddHealth function. As this cannot be overwritten in a blueprint, I have added a AddHealthBP function to JEnemyBP_Bar that just calls the C++ class. I updated the ArrowBP and JKnight to use this new function (nothing else was changed)

Using the above fix, I was able to overload Zorrander's "damage taking" to keep track of damage taken over time so we can cancel his smash upon enough damage taken.

I also made Growing Damage Zone increase size in a slight curve so it is more dangerous near the beginning and tamer as it gets bigger.