added a staggered state to the CPPs
Staggered is a different state than stunned.
how long being staggered lasts is based on an input to the AddHealth function. So it can be different for every attack (defaults to none).
Also behaves differently to account for if the knight if blocking (stun time is halved, blocking can be maintained)